8234274: [BACKOUT] JDK-8204128 NMT might report incorrect numbers for Compiler area

+++ src/hotspot/share/memory/arena.cpp
@@ -325,7 +325,7 @@ void Arena::destruct_contents() {
 // change the size
 void Arena::set_size_in_bytes(size_t size) {
   if (_size_in_bytes != size) {
-    ssize_t delta = size - size_in_bytes();
+    long delta = (long)(size - size_in_bytes());
     _size_in_bytes = size;
     MemTracker::record_arena_size_change(delta, _flags);
   }
